+++ app/Model/User.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace App\Models;
+namespace App\Model;
 
 use Illuminate\Foundation\Auth\User as Authenticatable;
 
 class User extends Authenticatable
 {
     protected $table = 'users';
+
     /**
       * Constant admin team
       * Set role admin 1
@@ -17,21 +18,23 @@ class User extends Authenticatable
     const ADMIN_TEAM_NAME = 'PHP';
     const ROLE_ADMIN = 1;
     const ROLE_USER = 0;
+
     /**
      * The attributes that are mass assignable.
      *
      * @var array
      */
     protected $fillable = [
-       'employ_code',
-       'name',
-       'email',
-       'team',
-       'avatar_url',
-       'is_admin',
-       'access_token',
-       'expires_at',
+     'employ_code',
+     'name',
+     'email',
+     'team',
+     'avatar_url',
+     'is_admin',
+     'access_token',
+     'expires_at',
     ];
+
      /**
       * Check admin
       *
